Removed required LDAP config for "
* Please reference #80 on the main project page - https://github.com/linkedin/oncall/issues/80 for some additional conversation on this PR.

The ldap_sync.py script is expecting an LDAP attribute "image_url" to exist for each object polled. If the attribute doesn't exist the script fails and won't sync user account data.

I recommend to remove "image_url" as a requirement as not everyone will have this attribute in their own LDAP/AD environments.
The result of removing "image_url" is that the value will simply remain blank in the OnCall database (an acceptable result in my opinion).